Image.magick() now calls im_magick2vips() directly
We were relying on libvips file type sniffing to trigger magick loads.
However, this did not work for image files which used libvips-supported
formats as wrappers for file types not supported by libvips.

For example, .3fr raw camera files are a tiff file containg only a
thumbnail, plus a huge binary data blob with the raw pixel data. libvips
will open this file with its built-in tiff reader (and see only the
thumbnail), when it should probably use ImageMagick to get the
reconstructed pixel data.

With this patch, Image.new() on a .3fr will get the thumbnail,
Image.magick() will reconstruct the pixels.

Thanks to cgat for the report, see:

https://github.com/jcupitt/ruby-vips/issues/30