Update README.md
Homebrew recently booted a bunch of Formulae to their science repository, VIPS has been included in the move. 

I opened up a ticket about this with Homebrew as the change has frustrated me a great deal and I think makes it substantially less easy to use.

Regardless, for the mean time, this is how to install VIPS via homebrew now.