add vips_thread_shutdown()
vips has some thread-private state and this needs to be cleaned up on
thread exit. Unfortunately it's not possible to do this automatically on
threads which vips didn't create, so for "foreign" threads, you need to
arrange to call vips_thread_shutdown() on thread termination.

this patch bumps the version number and adds VIPS::thread_shutdown

see

https://github.com/jcupitt/ruby-vips/issues/55