Fix race-condition during vips load
This avoids the problematic lazy-loading of the GI module which goes
horribly wrong if it is attempted by two threads at once.

If one really wants to delay laoding of the gi module, it is possible to
set $vips_skip_autoload = true before requiring vips or simply don't
require vips until it is needed.