Fix segmentation fault in Image.new_from_buffer
When string is in a binary encoding, then String#length always equals
String#bytesize. However, in other encodings like UTF-8 String#length
will be lower than String#bytesize if the string contains printable
multibyte characters (which will always happen with binary files like
images).

If we call Image.new_from_buffer with content of a binary file like an
image in UTF-8 encoding, it will raise a segmentation fault. This is
because it's communicating the buffer size to libvips incorrectly, due
to using String#length. We fix that by switching to String#bytesize.