Allow Operation options to have nil values
When you're passing an optional argument, sometimes it's convenient to
be able to pass nil, and expect the same behaviour as if you didn't pass
the option at all (have libvips use the default value).

  image.embed(background: nil)

Currently a type mismatch exception is raised if nil is passed for an
Operation option value, and this commit modifies resolving arguments in
a way that options with nil values are skipped and not forwarded to
libvips.